+++ CHANGELOG.rst
@@ -1,13 +1,15 @@
-0.99.0 (2023-08-22)
+0.99.0 (2023-08-23)
 -------------------
 
 New features
 ============
 
 * Add function modeling.set_reference_values
+* Add function modeling.set_lloq_data
 * Parse IV+oral models using CMT column
 * Specify DV in RUVSearch
 * Option to add logit IIV in ``add_iiv``
+* New options for remove_loq_data
 
 Changes
 =======